+++ sn_faucet/CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.3](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.4.2...sn_faucet-v0.4.3) - 2024-03-28
+
+### Other
+- updated the following local packages: sn_client
+
+## [0.4.2](https://github.com/joshuef/safe_network/compare/sn_faucet-v0.4.1...sn_faucet-v0.4.2) - 2024-03-28
+
+### Fixed
+- *(faucet)* bind to wan
+
+## [0.4.1](https://github.com/joshuef/safe_network/compare/sn_faucet-v0.4.0...sn_faucet-v0.4.1) - 2024-03-28
+
+### Fixed
+- *(faucet)* add build info)
+
 ## [0.4.0](https://github.com/joshuef/safe_network/compare/sn_faucet-v0.3.85...sn_faucet-v0.4.0) - 2024-03-27
 
 ### Added